requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
timm>=0.9.0
albumentations>=1.3.0
albumentations-pytorch>=0.1.0
torchmetrics>=0.10.0
pandas>=1.5.0
spacy>=3.0
matplotlib>=3.5.0
Pillow>=9.0.0
pyyaml>=6.0.0
